Re-construct folder 2
