--- GameScene.java
@@ -1,14 +1,0 @@
-package Scenes;
-
-import main.Game;
-
-public class GameScene {
-    private Game game;
-    public GameScene(Game game){
-        this.game = game;
-    }
-
-    public Game getGame() {
-        return game;
-    }
-}
