+++ Scene/Settings.java
@@ -1,0 +1,68 @@
+package Scenes;
+
+import UI.MyButton;
+import main.Game;
+
+import java.awt.*;
+
+import static main.GameStates.MENU;
+import static main.GameStates.SetGameState;
+
+public class Settings extends GameScene implements SceneMethods{
+    private MyButton bMenu;
+
+    public Settings(Game game) {
+        super(game);
+        initButtons();
+    }
+
+    private void initButtons() {
+        bMenu = new MyButton("Menu", 2, 2, 100, 30);
+    }
+
+    @Override
+    public void render(Graphics g) {
+        g.setColor(Color.YELLOW);
+        g.fillRect(0, 0, 640, 640);
+
+        drawButtons(g);
+    }
+
+    private void drawButtons(Graphics g) {
+        bMenu.draw(g);
+    }
+
+    @Override
+    public void mouseClicked(int x, int y) {
+        if (bMenu.getBounds().contains(x, y))
+            SetGameState(MENU);
+
+    }
+
+    @Override
+    public void mouseMoved(int x, int y) {
+        bMenu.setMouseOver(false);
+        if (bMenu.getBounds().contains(x, y))
+            bMenu.setMouseOver(true);
+    }
+
+    @Override
+    public void mousePressed(int x, int y) {
+        if (bMenu.getBounds().contains(x, y))
+            bMenu.setMousePressed(true);
+    }
+
+    @Override
+    public void mouseReleased(int x, int y) {
+        resetButtons();
+    }
+
+    @Override
+    public void mouseDragged(int x, int y) {
+
+    }
+
+    private void resetButtons() {
+        bMenu.resetBooleans();
+    }
+}
