--- SceneMethods.java
@@ -1,12 +1,0 @@
-package Scenes;
-
-import java.awt.*;
-
-public interface SceneMethods {
-    public void render(Graphics g);
-    public void mouseClicked(int x , int y);
-    public void mouseMoved(int x, int y);
-    public void mousePressed(int x, int y);
-    public void mouseReleased(int x, int y);
-    public void mouseDragged(int x, int y);
-}
